requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Classful==0.14.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
paho-mqtt==1.5.0
Werkzeug==1.0.1